backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
PyJWT==2.10.1
Werkzeug==2.3.7
gunicorn==21.2.0
APScheduler==3.10.4
python-dateutil==2.8.2
